make aggregator proxy round aware
