+++ evm-contracts/test/v0.6/EACAggregatorProxy.test.ts
@@ -26,7 +26,7 @@ beforeAll(async () => {
   defaultAccount = users.roles.defaultAccount
 })
 
-describe('AccessControlledAggregatorProxy', () => {
+describe('EACAggregatorProxy', () => {
   const deposit = h.toWei('100')
   const answer = h.numToBytes32(54321)
   const answer2 = h.numToBytes32(54320)
@@ -40,6 +40,7 @@ describe('AccessControlledAggregatorProxy', () => {
   let aggregator: contract.Instance<MockV3AggregatorFactory>
   let aggregator2: contract.Instance<MockV3AggregatorFactory>
   let proxy: contract.Instance<EACAggregatorProxyFactory>
+  const epochBase = h.bigNum(4294967296)
 
   const deployment = setup.snapshot(provider, async () => {
     link = await linkTokenFactory.connect(defaultAccount).deploy()
@@ -65,6 +66,8 @@ describe('AccessControlledAggregatorProxy', () => {
       'confirmAggregator',
       'decimals',
       'description',
+      'epoch',
+      'epochAggregators',
       'getAnswer',
       'getRoundData',
       'getTimestamp',
@@ -95,11 +98,11 @@ describe('AccessControlledAggregatorProxy', () => {
     })
 
     it('#getAnswer', async () => {
-      await proxy.connect(personas.Carol).getAnswer(1)
+      await proxy.connect(personas.Carol).getAnswer(epochBase.add(1))
     })
 
     it('#getTimestamp', async () => {
-      await proxy.connect(personas.Carol).getTimestamp(1)
+      await proxy.connect(personas.Carol).getTimestamp(epochBase.add(1))
     })
 
     it('#latestRound', async () => {
@@ -107,7 +110,7 @@ describe('AccessControlledAggregatorProxy', () => {
     })
 
     it('#getRoundData', async () => {
-      await proxy.connect(personas.Carol).getRoundData(1)
+      await proxy.connect(personas.Carol).getRoundData(epochBase.add(1))
     })
 
     it('#proposedGetRoundData', async () => {
@@ -159,11 +162,9 @@ describe('AccessControlledAggregatorProxy', () => {
     })
 
     it('getRoundData works', async () => {
-      const latestRound = await proxy.latestRound()
-      await proxy.latestRound()
-      const round = await proxy.getRoundData(latestRound)
-      await proxy.getRoundData(latestRound)
-      matchers.bigNum(roundId, round.roundId)
+      const proxyRoundId = await proxy.latestRound()
+      const round = await proxy.getRoundData(proxyRoundId)
+      matchers.bigNum(proxyRoundId, round.roundId)
       matchers.bigNum(answer, round.answer)
       matchers.bigNum(startedAt, round.startedAt)
       matchers.bigNum(timestamp, round.updatedAt)
