--- evm-contracts/test/v0.6/RoundAwareAggregatorProxy.test.ts
@@ -1,187 +1,0 @@
-import {
-  contract,
-  helpers as h,
-  matchers,
-  setup,
-} from '@chainlink/test-helpers'
-import { assert } from 'chai'
-import { ethers } from 'ethers'
-import { BigNumber } from 'ethers/utils'
-import { MockV3AggregatorFactory } from '../../ethers/v0.6/MockV3AggregatorFactory'
-import { RoundAwareAggregatorProxyFactory } from '../../ethers/v0.6/RoundAwareAggregatorProxyFactory'
-
-let personas: setup.Personas
-let defaultAccount: ethers.Wallet
-
-const provider = setup.provider()
-const linkTokenFactory = new contract.LinkTokenFactory()
-const aggregatorFactory = new MockV3AggregatorFactory()
-const aggregatorProxyFactory = new RoundAwareAggregatorProxyFactory()
-
-beforeAll(async () => {
-  const users = await setup.users(provider)
-
-  personas = users.personas
-  defaultAccount = users.roles.defaultAccount
-})
-
-describe('AggregatorProxy', () => {
-  const deposit = h.toWei('100')
-  const response = h.numToBytes32(54321)
-  const response2 = h.numToBytes32(67890)
-  const decimals = 18
-  const epochBase = h.bigNum(4294967296)
-
-  let link: contract.Instance<contract.LinkTokenFactory>
-  let aggregator: contract.Instance<MockV3AggregatorFactory>
-  let aggregator2: contract.Instance<MockV3AggregatorFactory>
-  let proxy: contract.Instance<RoundAwareAggregatorProxyFactory>
-  const deployment = setup.snapshot(provider, async () => {
-    link = await linkTokenFactory.connect(defaultAccount).deploy()
-    aggregator = await aggregatorFactory
-      .connect(defaultAccount)
-      .deploy(decimals, response)
-    await link.transfer(aggregator.address, deposit)
-    proxy = await aggregatorProxyFactory
-      .connect(defaultAccount)
-      .deploy(aggregator.address)
-  })
-
-  beforeEach(async () => {
-    await deployment()
-  })
-
-  it('has a limited public interface', () => {
-    matchers.publicAbi(aggregatorProxyFactory, [
-      'aggregator',
-      'confirmAggregator',
-      'decimals',
-      'description',
-      'epoch',
-      'epochAggregators',
-      'getAnswer',
-      'getRoundData',
-      'getTimestamp',
-      'latestAnswer',
-      'latestRound',
-      'latestRoundData',
-      'latestTimestamp',
-      'version',
-      'proposeAggregator',
-      'proposedAggregator',
-      'proposedGetRoundData',
-      'proposedLatestRoundData',
-      // Ownable methods:
-      'acceptOwnership',
-      'owner',
-      'transferOwnership',
-    ])
-  })
-
-  describe('constructor', () => {
-    it('sets the proxy epoch and aggregator', async () => {
-      matchers.bigNum(1, await proxy.epoch())
-      assert.equal(aggregator.address, await proxy.epochAggregators(1))
-    })
-  })
-
-  describe('#latestRound', () => {
-    it('pulls the rate from the aggregator', async () => {
-      matchers.bigNum(epochBase.add(1), await proxy.latestRound())
-    })
-  })
-
-  describe('#latestAnswer', () => {
-    it('pulls the rate from the aggregator', async () => {
-      matchers.bigNum(response, await proxy.latestAnswer())
-      const latestRound = await proxy.latestRound()
-      matchers.bigNum(response, await proxy.getAnswer(latestRound))
-    })
-
-    describe('after being updated to another contract', () => {
-      let preUpdateRoundId: BigNumber
-      let preUpdateAnswer: BigNumber
-
-      beforeEach(async () => {
-        preUpdateRoundId = await proxy.latestRound()
-        preUpdateAnswer = await proxy.latestAnswer()
-
-        aggregator2 = await aggregatorFactory
-          .connect(defaultAccount)
-          .deploy(decimals, response2)
-        await link.transfer(aggregator2.address, deposit)
-        matchers.bigNum(response2, await aggregator2.latestAnswer())
-
-        await proxy.proposeAggregator(aggregator2.address)
-        await proxy.confirmAggregator(aggregator2.address)
-      })
-
-      it('pulls the rate from the new aggregator', async () => {
-        matchers.bigNum(response2, await proxy.latestAnswer())
-        const latestRound = await proxy.latestRound()
-        matchers.bigNum(response2, await proxy.getAnswer(latestRound))
-      })
-
-      it('allows requests of to previous aggregators', async () => {
-        matchers.bigNum(
-          preUpdateAnswer,
-          await proxy.getAnswer(preUpdateRoundId),
-        )
-      })
-    })
-  })
-
-  describe('#confirmAggregator', () => {
-    beforeEach(async () => {
-      await proxy.transferOwnership(personas.Carol.address)
-      await proxy.connect(personas.Carol).acceptOwnership()
-
-      aggregator2 = await aggregatorFactory
-        .connect(defaultAccount)
-        .deploy(decimals, 1)
-
-      assert.equal(aggregator.address, await proxy.aggregator())
-    })
-
-    describe('when called by the owner', () => {
-      beforeEach(async () => {
-        await proxy
-          .connect(personas.Carol)
-          .proposeAggregator(aggregator2.address)
-      })
-
-      it('increases the epoch', async () => {
-        matchers.bigNum(1, await proxy.epoch())
-
-        await proxy
-          .connect(personas.Carol)
-          .confirmAggregator(aggregator2.address)
-
-        matchers.bigNum(2, await proxy.epoch())
-      })
-
-      it('increases the round ID', async () => {
-        matchers.bigNum(epochBase.add(1), await proxy.latestRound())
-
-        await proxy
-          .connect(personas.Carol)
-          .confirmAggregator(aggregator2.address)
-
-        matchers.bigNum(epochBase.mul(2).add(1), await proxy.latestRound())
-      })
-
-      it('sets the proxy epoch and aggregator', async () => {
-        assert.equal(
-          '0x0000000000000000000000000000000000000000',
-          await proxy.epochAggregators(2),
-        )
-
-        await proxy
-          .connect(personas.Carol)
-          .confirmAggregator(aggregator2.address)
-
-        assert.equal(aggregator2.address, await proxy.epochAggregators(2))
-      })
-    })
-  })
-})
